+++ src/main/java/com/redbeemedia/enigma/core/player/listener/IEnigmaPlayerListener.java
@@ -23,4 +23,6 @@ public interface IEnigmaPlayerListener extends IInternalListener {
     void onPlaybackSessionChanged(IPlaybackSession from, IPlaybackSession to);
 
     void onProgramChanged(IProgram from, IProgram to);
+
+    void sendPlaybackStartedEvent();
 }