+++ CHANGELOG.md
@@ -31,6 +31,8 @@ DevEnigmaRiverAndroidCore
 DevEnigmaRiverAndroidExoPlayerIntegration
 DevEnigmaRiverAndroidDownload
 `
+* EMP-18002- 10min inactivity issue
+* EMP-18242- Concurrent Stream issue
 * EMP-18125- Add support for custom key
 * EMP-18196- Added API for the expiration date